server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Simple gRPC server for embeddings."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "sentence-transformers>=4.0.2",
]
